ENH: Modified the SetRadiusXYZ, the Radius can be floating point
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=3620
git-svn-id: http://svn.slicer.org/Slicer4/trunk@3620 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/vtkMRMLROIListNode.cxx
@@ -659,7 +659,7 @@ int vtkMRMLROIListNode::SetNthROIRadiusIJK(int n, float Radiusi, float Radiusj, 
     if (VolumeNode)
       {
       double* spacing = VolumeNode->GetSpacing();
-      node->SetRadiusXYZ((int)Radiusi * spacing[0], (int)Radiusj * spacing[1], (int)Radiusk * spacing[2]);
+      node->SetRadiusXYZ(Radiusi * spacing[0], Radiusj * spacing[1], Radiusk * spacing[2]);
       }
     else
       {
